Split code into separate files (one class per file)
