+++ src/SDKs/Azure.Core/data-plane/Azure.Core/Attributes/UsageErrorsAttribute.cs
@@ -6,31 +6,6 @@ using System;
 
 namespace Azure.Core
 {
-    public class KnownExceptionAttribute : Attribute
-    {
-        public KnownExceptionAttribute(Type exceptionType)
-            => ExceptionType = exceptionType;
-
-        public Type ExceptionType { get; }
-
-        public string Message { get; set; }
-    }
-
-    /// <summary>
-    /// Represents errors that the application might handle.
-    /// </summary>
-    [AttributeUsage(AttributeTargets.Method, AllowMultiple = true)]
-    public class HttpErrorAttribute : KnownExceptionAttribute
-    {
-        public HttpErrorAttribute(Type exceptionType, int statusCode)
-            : base(exceptionType)
-        {
-            StatusCode = statusCode;
-        }
-
-        public int StatusCode { get; }
-    }
-
     /// <summary>
     /// Represents errors resulting from misuse of the API. The application code should be changed/fixed.
     /// </summary>
