+++ src/SDKs/Azure.Core/data-plane/Azure.Core/Net/Pipeline/FixedPolicy.cs
@@ -1,0 +1,32 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for
+// license information.
+
+using System;
+
+namespace Azure.Core.Http.Pipeline
+{
+    class FixedPolicy : RetryPolicy {
+        int _maxRetries;
+        TimeSpan _delay;
+        int[] _retriableCodes;
+
+        public FixedPolicy(int[] retriableCodes, int maxRetries, TimeSpan delay)
+        {
+            if (retriableCodes == null) throw new ArgumentNullException(nameof(retriableCodes));
+
+            _maxRetries = maxRetries;
+            _delay = delay;
+            _retriableCodes = retriableCodes;
+            Array.Sort(_retriableCodes);
+        }
+
+        protected override bool ShouldRetry(HttpMessage message, int attempted, out TimeSpan delay)
+        {
+            delay = _delay;
+            if (attempted > _maxRetries) return false;
+            if(Array.BinarySearch(_retriableCodes, message.Response.Status) < 0) return false;
+            return true;
+        }
+    }
+}
