+++ src/SDKs/Azure.Core/data-plane/Azure.Core/Net/Pipeline/PipelinePolicy.cs
@@ -23,17 +23,4 @@ namespace Azure.Core.Http.Pipeline
 
         protected AzureEventSource Log = AzureEventSource.Singleton;
     }
-
-    public abstract class PipelineTransport : PipelinePolicy
-    {
-        public abstract Task ProcessAsync(HttpMessage message);
-
-        public abstract HttpMessage CreateMessage(PipelineOptions options, CancellationToken cancellation);
-
-        public sealed override async Task ProcessAsync(HttpMessage message, ReadOnlyMemory<PipelinePolicy> next)
-        {
-            if (next.Length == 0) await ProcessAsync(message).ConfigureAwait(false);
-            else throw new ArgumentOutOfRangeException(nameof(next));
-        }
-    }
 }
